(split)bug:60273 Don't update HighestSubmittedToIndex by non-reindex thread during reindex
http://bugzilla.zimbra.com/show_bug.cgi?id=60273

Copied from Perforce
 Change: 316987

+++ src/java/com/zimbra/cs/mailbox/IndexHelper.java
@@ -947,6 +947,12 @@ public class IndexHelper {
     }
 
     void indexingPartOfEndTransaction(List<IndexItemEntry> itemsToIndex, List<Integer> itemsToDelete) {
+        // if reindex is in progress, don't update HighestSubmittedToIndex.
+        if (mFullReindexInProgress && !Thread.currentThread().getName().startsWith(sReIndexThreadPool.getName())) {
+            ZimbraLog.index.warn("Reindex is in progress");
+            return;
+        }
+
         if (itemsToDelete != null && !itemsToDelete.isEmpty()) {
             try {
                 getMailboxIndex().deleteDocuments(itemsToDelete);
